requirements.txt
python=3.10.20
pytorch=2.5.1
transformers=5.3.0
accelerate=1.13.0